+++ app/controllers/beyond_canvas/application_controller.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 module BeyondCanvas
-  class ApplicationController < ActionController::Base
+  class ApplicationController < ActionController::Base # :nodoc:
     protect_from_forgery with: :exception
 
     include ::BeyondCanvas::StatusCodes