+++ app/form_builders/beyond_canvas/form_builder.rb
@@ -1,9 +1,9 @@
 # frozen_string_literal: true
 
 module BeyondCanvas
-  class FormBuilder < ActionView::Helpers::FormBuilder
+  class FormBuilder < ActionView::Helpers::FormBuilder # :nodoc:
     def field_wrapper(attribute, args, &block)
-      label = args[:label].present? ? args[:label] : attribute.to_s.humanize
+      label = args[:label].presence || attribute.to_s.humanize
 
       errors = object.errors[attribute].join(', ') if object.respond_to?(:errors) && object.errors.include?(attribute)
 
@@ -11,13 +11,13 @@ module BeyondCanvas
         @template.content_tag(:label, label, class: 'input__label') +
           @template.content_tag(:div, class: 'relative') do
             block.call +
-              (@template.content_tag(:label, errors, class: 'input__error') unless errors.blank?)
+              (@template.content_tag(:label, errors, class: 'input__error') if errors.present?)
           end +
           (@template.content_tag(:div, args[:hint].html_safe, class: 'input__hint') if args[:hint].present?)
       end
     end
 
-    [:email, :text, :number, :password].each do |method|
+    %i[email text number password].each do |method|
       define_method :"#{method}_field" do |attribute, args = {}|
         field_wrapper(attribute, args) do
           super(attribute, args)
@@ -27,7 +27,7 @@ module BeyondCanvas
 
     def file_field(attribute, args = {})
       field_wrapper(attribute, args) do
-        filed_identifyer = "#{attribute}_#{(Time.now.to_f * 1000).to_i.to_s}"
+        filed_identifyer = "#{attribute}_#{(Time.now.to_f * 1000).to_i}"
 
         args.merge!(id: filed_identifyer)
             .merge!(hidden: true)
@@ -37,10 +37,14 @@ module BeyondCanvas
 
         @template.content_tag(:div, class: 'input__file') do
           super(attribute, args) +
-          @template.content_tag(:label, for: filed_identifyer, class: 'input__file__label button__transparent--primary') do
-            args[:data][:button_text] || 'Choose file'
-          end +
-          @template.content_tag(:span, args[:data][:no_file_text] || 'No file chosen', class: "input__file__text #{filed_identifyer}")
+            @template.content_tag(:label,
+                                  for: filed_identifyer,
+                                  class: 'input__file__label button__transparent--primary') do
+              args[:data][:button_text] || 'Choose file'
+            end +
+            @template.content_tag(:span,
+                                  args[:data][:no_file_text] || 'No file chosen',
+                                  class: "input__file__text #{filed_identifyer}")
         end
       end
     end