+++ app/javascript/beyond_canvas/base.js
@@ -1,0 +1,6 @@
+import 'jquery'
+import 'jquery-ujs'
+
+import './initializers/buttons'
+import './initializers/flash'
+import './initializers/inputs'