+++ lib/beyond_canvas/asset_registration.rb
@@ -1,0 +1,29 @@
+# frozen_string_literal: true
+
+module BeyondCanvas
+  module AssetRegistration # :nodoc:
+    def register_stylesheet(path, options = {})
+      stylesheets[path] = options
+    end
+
+    def stylesheets
+      @stylesheets ||= {}
+    end
+
+    def clear_stylesheets!
+      stylesheets.clear
+    end
+
+    def register_javascript(name)
+      javascripts.add name
+    end
+
+    def javascripts
+      @javascripts ||= Set.new
+    end
+
+    def clear_javascripts!
+      javascripts.clear
+    end
+  end
+end