+++ lib/generators/beyond_canvas/custom_styles/custom_styles_generator.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+module BeyondCanvas
+  module Generators
+    class CustomStylesGenerator < Rails::Generators::Base # :nodoc:
+      source_root File.expand_path('templates', __dir__)
+
+      def copy_initializer
+        template 'beyond_canvas_custom_styles.sass', 'app/assets/stylesheets/_beyond_canvas_custom_styles.sass'
+      end
+    end
+  end
+end