+++ CHANGELOG.md
@@ -1,3 +1,23 @@
+### v0.14.0.pre
+
+* deprecations
+  * `application.css` is no longer loaded by Beyond Canvas
+  * Move stylesheets from `.sass` to `.scss`
+  * Move partials from `.html.slim` to `.html.erb`
+  * Deprecations on `BeyondCanvas` configuration file:
+    * `:public_logo` is substituted by `:site_logo`
+
+* features
+  * Add support for both Webpacker and Sprockets
+  * `BeyondCanvas` configuration:
+    * Added support for `:favicon`
+    * Added support for `:site_title`
+    * Added the possibility to register stylesheets and javascripts
+
+* enhancements
+  * Remove Font Awesome dependency (now `.svg` icons are used)
+  * Title is now taken from `BeyondCanvas` initializer
+
 ### v0.13.1.pre
 
 * bug-fixes